Move some ISemaphore tests
(cherry picked from commit d3e47dea230a2574270027f691769f7c6bbca77f)
